Implement future return values via callback
Implement the future return value handling via callback handling,
roughly as it is done in Boost.Asio.

Rework the task abstraction to invoke the stored callback upon poll
completion. This takes some gymnastics with abstract base classes
inherited by genericised child classes.

Also, make the FfiPoll and FfiFuture struct generic. The compiler is not
happy about an extern "C" function returning a generic struct, but at
least with GCC it works for now. The alternative would be to define
separate structs for each result type, which is exactly what templates
were invented for.

Signed-off-by: Florian Limberger <florian.limberger@inovex.de>